VPN-3508 - part 14 - Cleanup networkRequests
In this commit I do the following:
- implement a callback system to overwrite the default network behavior (unit-test, wasm)
- move the single network request constructors in the corresponding code
- move networkrequest.{cpp,h} into /src/shared
- remove mocnetworkrequest from unit-tests

Revising the network-request implementation I also found a bug (and fixed it) in
the benchmark task logic where, if the upload is slower than the download, we
do not show the proper error message.
