+++ src/apps/vpn/commands/commandui.cpp
@@ -36,6 +36,7 @@
 #include "models/featuremodel.h"
 #include "models/recentconnections.h"
 #include "mozillavpn.h"
+#include "networkrequest.h"
 #include "notificationhandler.h"
 #include "productshandler.h"
 #include "purchasehandler.h"
@@ -86,6 +87,7 @@
 #endif
 
 #ifdef MZ_WASM
+#  include "platforms/wasm/wasmnetworkrequest.h"
 #  include "platforms/wasm/wasmwindowcontroller.h"
 #endif
 
@@ -664,6 +666,11 @@ int CommandUI::run(QStringList& tokens) {
 
 #ifdef MZ_WASM
     WasmWindowController wasmWindowController;
+
+    NetworkRequest::setRequestHandler(WasmNetworkRequest::deleteRequest,
+                                      WasmNetworkRequest::getRequest,
+                                      WasmNetworkRequest::postRequest,
+                                      WasmNetworkRequest::postRequestIODevice);
 #endif
 
 #ifdef MVPN_WEBEXTENSION
