+++ src/apps/vpn/connectionbenchmark/benchmarktasktransfer.cpp
@@ -74,7 +74,8 @@ void BenchmarkTaskTransfer::createNetworkRequest() {
   NetworkRequest* request = nullptr;
   switch (m_type) {
     case BenchmarkDownload: {
-      request = NetworkRequest::createForGetUrl(this, m_url.toString());
+      request = NetworkRequest::create(this);
+      request->get(m_url);
       break;
     }
     case BenchmarkUpload: {
@@ -85,8 +86,10 @@ void BenchmarkTaskTransfer::createNetworkRequest() {
         emit finished(0, true);
         emit completed();
       };
-      request = NetworkRequest::createForUploadData(this, m_url.toString(),
-                                                    uploadData);
+      request = NetworkRequest::create(this, 200);
+      request->requestInternal().setHeader(QNetworkRequest::ContentTypeHeader,
+                                           "application/x-www-form-urlencoded");
+      request->post(m_url, uploadData);
       break;
     }
   }
@@ -102,8 +105,21 @@ void BenchmarkTaskTransfer::createNetworkRequestWithRecord(
   NetworkRequest* request = nullptr;
   switch (m_type) {
     case BenchmarkDownload: {
-      request = NetworkRequest::createForGetHostAddress(this, m_url.toString(),
-                                                        record.value());
+      QUrl requestUrl(m_url);
+      QString hostname = requestUrl.host();
+
+      // Rewrite the request URL to use an explicit host address.
+      if (record.value().protocol() == QAbstractSocket::IPv6Protocol) {
+        requestUrl.setHost("[" + record.value().toString() + "]");
+      } else {
+        requestUrl.setHost(record.value().toString());
+      }
+
+      request = NetworkRequest::create(this, 200);
+      request->requestInternal().setRawHeader("Host", hostname.toLocal8Bit());
+      request->requestInternal().setPeerVerifyName(hostname);
+
+      request->get(requestUrl);
       break;
     }
     case BenchmarkUpload: {
@@ -114,8 +130,24 @@ void BenchmarkTaskTransfer::createNetworkRequestWithRecord(
         emit finished(0, true);
         emit completed();
       };
-      request = NetworkRequest::createForUploadDataHostAddress(
-          this, m_url.toString(), uploadData, record.value());
+      QUrl requestUrl(m_url);
+      QString hostname = requestUrl.host();
+
+      // Rewrite the request URL to use an explicit host address.
+      if (record.value().protocol() == QAbstractSocket::IPv6Protocol) {
+        requestUrl.setHost("[" + record.value().toString() + "]");
+      } else {
+        requestUrl.setHost(record.value().toString());
+      }
+
+      request = NetworkRequest::create(this, 200);
+      request->requestInternal().setHeader(QNetworkRequest::ContentTypeHeader,
+                                           "application/x-www-form-urlencoded");
+
+      request->requestInternal().setRawHeader("Host", hostname.toLocal8Bit());
+      request->requestInternal().setPeerVerifyName(hostname);
+
+      request->post(requestUrl, uploadData);
       break;
     }
     default: {
