+++ src/apps/vpn/tasks/account/taskaccount.cpp
@@ -22,7 +22,10 @@ TaskAccount::TaskAccount(
 TaskAccount::~TaskAccount() { MZ_COUNT_DTOR(TaskAccount); }
 
 void TaskAccount::run() {
-  NetworkRequest* request = NetworkRequest::createForAccount(this);
+  NetworkRequest* request = NetworkRequest::create(this, 200);
+  request->auth(MozillaVPN::authorizationHeader());
+  request->get(
+      QString("%1/api/v1/vpn/account").arg(AppConstants::apiBaseUrl()));
 
   connect(request, &NetworkRequest::requestFailed, this,
           [this](QNetworkReply::NetworkError error, const QByteArray&) {
