+++ src/apps/vpn/tasks/addon/taskaddon.cpp
@@ -21,10 +21,9 @@ TaskAddon::TaskAddon(const QString& addonId, const QByteArray& sha256)
 TaskAddon::~TaskAddon() { MZ_COUNT_DTOR(TaskAddon); }
 
 void TaskAddon::run() {
-  NetworkRequest* request = NetworkRequest::createForGetUrl(
-      this,
-      QString("%1%2.rcc").arg(AddonManager::addonServerAddress(), m_addonId),
-      200);
+  NetworkRequest* request = NetworkRequest::create(this, 200);
+  request->get(
+      QString("%1%2.rcc").arg(AddonManager::addonServerAddress(), m_addonId));
 
   connect(request, &NetworkRequest::requestFailed, this,
           [this](QNetworkReply::NetworkError error, const QByteArray&) {
