+++ src/apps/vpn/tasks/getfeaturelist/taskgetfeaturelist.cpp
@@ -7,6 +7,7 @@
 #include <QJsonDocument>
 #include <QJsonObject>
 
+#include "appconstants.h"
 #include "errorhandler.h"
 #include "leakdetector.h"
 #include "logger.h"
@@ -24,7 +25,9 @@ TaskGetFeatureList::TaskGetFeatureList() : Task("TaskGetFeatureList") {
 TaskGetFeatureList::~TaskGetFeatureList() { MZ_COUNT_DTOR(TaskGetFeatureList); }
 
 void TaskGetFeatureList::run() {
-  NetworkRequest* request = NetworkRequest::createForGetFeatureList(this);
+  NetworkRequest* request = NetworkRequest::create(this, 200);
+  request->get(
+      QString("%1/api/v1/vpn/featurelist").arg(AppConstants::apiBaseUrl()));
 
   connect(request, &NetworkRequest::requestFailed, this,
           [this](QNetworkReply::NetworkError error, const QByteArray&) {
