+++ src/apps/vpn/tasks/getsubscriptiondetails/taskgetsubscriptiondetails.cpp
@@ -4,6 +4,9 @@
 
 #include "taskgetsubscriptiondetails.h"
 
+#include <QJsonObject>
+
+#include "appconstants.h"
 #include "authenticationinapp/authenticationinapp.h"
 #include "authenticationinapp/authenticationinappsession.h"
 #include "authenticationlistener.h"
@@ -44,8 +47,10 @@ void TaskGetSubscriptionDetails::run() {
 }
 
 void TaskGetSubscriptionDetails::runInternal() {
-  NetworkRequest* request =
-      NetworkRequest::createForGetSubscriptionDetails(this);
+  NetworkRequest* request = NetworkRequest::create(this, 200);
+  request->auth(MozillaVPN::authorizationHeader());
+  request->get(QString("%1/api/v1/vpn/subscriptionDetails")
+                   .arg(AppConstants::apiBaseUrl()));
 
   connect(
       request, &NetworkRequest::requestFailed, this,
@@ -139,9 +144,12 @@ void TaskGetSubscriptionDetails::initAuthentication() {
             logger.debug() << "Authentication completed with code:"
                            << logger.sensitive(pkceCodeSuccess);
 
-            NetworkRequest* request =
-                NetworkRequest::createForAuthenticationVerification(
-                    this, pkceCodeSuccess, pkceCodeVerifier);
+            NetworkRequest* request = NetworkRequest::create(this, 200);
+            request->post(
+                QString("%1/api/v2/vpn/login/verify")
+                    .arg(AppConstants::apiBaseUrl()),
+                QJsonObject{{"code", pkceCodeSuccess},
+                            {"code_verifier", QString(pkceCodeVerifier)}});
 
             connect(
                 request, &NetworkRequest::requestFailed, this,
@@ -176,7 +184,8 @@ void TaskGetSubscriptionDetails::initAuthentication() {
                 break;
 
               default:
-                // All the other states should be handled by the front-end code.
+                // All the other states should be handled by the
+                // front-end code.
                 break;
             }
           });
