+++ src/apps/vpn/tasks/removedevice/taskremovedevice.cpp
@@ -36,8 +36,10 @@ void TaskRemoveDevice::run() {
   logger.debug() << "Removing the device with public key"
                  << logger.keys(m_publicKey);
 
-  NetworkRequest* request =
-      NetworkRequest::createForDeviceRemoval(this, m_publicKey);
+  NetworkRequest* request = NetworkRequest::create(this, 204);
+  request->deleteRequest(QString("%1/api/v1/vpn/device/%2")
+                             .arg(AppConstants::apiBaseUrl(),
+                                  QUrl::toPercentEncoding(m_publicKey)));
 
   connect(request, &NetworkRequest::requestFailed, this,
           [this](QNetworkReply::NetworkError error, const QByteArray&) {
