+++ src/apps/vpn/tasks/sentry/tasksentry.cpp
@@ -76,7 +76,13 @@ void TaskSentry::run() {
 }
 
 void TaskSentry::sendRequest() {
-  NetworkRequest* request = NetworkRequest::createForSentry(this, m_envelope);
+  NetworkRequest* request = NetworkRequest::create(this, 200);
+
+  request->requestInternal().setHeader(QNetworkRequest::ContentTypeHeader,
+                                       "application/x-sentry-envelope");
+  request->requestInternal().setRawHeader("dsn",
+                                          AppConstants::SENTRY_DSN_ENDPOINT);
+  request->post(QUrl(AppConstants::SENTRY_ENVELOPE_INGESTION), m_envelope);
 
   connect(request, &NetworkRequest::requestFailed, this,
           [this](QNetworkReply::NetworkError error, const QByteArray&) {
