+++ src/apps/vpn/tasks/servers/taskservers.cpp
@@ -23,7 +23,10 @@ TaskServers::TaskServers(
 TaskServers::~TaskServers() { MZ_COUNT_DTOR(TaskServers); }
 
 void TaskServers::run() {
-  NetworkRequest* request = NetworkRequest::createForServers(this);
+  NetworkRequest* request = NetworkRequest::create(this, 200);
+  request->auth(MozillaVPN::authorizationHeader());
+  request->get(
+      QString("%1/api/v1/vpn/servers").arg(AppConstants::apiBaseUrl()));
 
   connect(request, &NetworkRequest::requestFailed, this,
           [this](QNetworkReply::NetworkError error, const QByteArray&) {
