+++ tests/unit/main.cpp
@@ -7,6 +7,7 @@
 #include "l18nstrings.h"
 #include "leakdetector.h"
 #include "loghandler.h"
+#include "networkrequest.h"
 #include "settingsholder.h"
 
 QVector<TestHelper::NetworkConfig> TestHelper::networkConfig;
@@ -29,6 +30,25 @@ QObject* TestHelper::findTest(const QString& name) {
 
 TestHelper::TestHelper() { testList.append(this); }
 
+// static
+void TestHelper::networkRequestGeneric(NetworkRequest* request) {
+  Q_ASSERT(!TestHelper::networkConfig.isEmpty());
+  TestHelper::NetworkConfig nc = TestHelper::networkConfig.takeFirst();
+
+  QTimer::singleShot(0, request, [request, nc]() {
+    request->deleteLater();
+
+    if (nc.m_status == TestHelper::NetworkConfig::Failure) {
+      emit request->requestFailed(
+          QNetworkReply::NetworkError::HostNotFoundError, "");
+    } else {
+      Q_ASSERT(nc.m_status == TestHelper::NetworkConfig::Success);
+
+      emit request->requestCompleted(nc.m_body);
+    }
+  });
+}
+
 int main(int argc, char* argv[]) {
 #ifdef MZ_DEBUG
   LeakDetector leakDetector;
@@ -46,6 +66,10 @@ int main(int argc, char* argv[]) {
   QCoreApplication app(argc, argv);
 
   int failures = 0;
+
+  NetworkRequest::setRequestHandler(
+      TestHelper::networkRequestDelete, TestHelper::networkRequestGet,
+      TestHelper::networkRequestPost, TestHelper::networkRequestPostIODevice);
 
   L18nStrings::initialize();
   LogHandler::enableStderr();
