+++ tests/unit/testipfinder.cpp
@@ -6,18 +6,21 @@
 
 #include <QHostInfo>
 
+#include "appconstants.h"
 #include "helper.h"
-#include "networkrequest.h"
 #include "settingsholder.h"
+#include "simplenetworkmanager.h"
 #include "tasks/ipfinder/taskipfinder.h"
 
 void TestIpFinder::ipv4AndIpv6() {
+  SimpleNetworkManager snm;
+
   SettingsHolder settingsHolder;
 
   TestHelper::networkConfig.clear();
 
   QEventLoop loop;
-  QUrl url(NetworkRequest::apiBaseUrl());
+  QUrl url(AppConstants::apiBaseUrl());
 
   bool ipv4Expected = false;
   bool ipv6Expected = false;
